exporter/requirements.txt
prometheus_client
smbus2
RPi.bme280
pms7003
